requirements.txt
Django==3.2.15
django-debug-toolbar==3.6.0
django-filter==22.1
djangorestframework==3.14.0
pytest-django==4.5.2
flake8
psycopg2==2.9.3
pytest-cov==4.0.0
black==22.10.0
requests==2.26.0